AnalisisCoberturaAPI/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary
python-dotenv==1.0.1
